+++ R/autospc.R
@@ -54,6 +54,10 @@
 #' ## SPC Parameters
 #' Parameters that control how centre line and control limits are established
 #' for each period, and details of how SPC rules are applied
+#' @param na_ends_run Boolean determining whether a missing point starts a new
+#' run, for the purpose of the shift rule. TRUE minimises the risk of a false
+#' positive shift rule break arising from missing data; FALSE minimises the risk
+#' of a false negative.
 #' @param max_exclusions The maximum number of extreme points to exclude from
 #' limit calculations.
 #' @param highlight_exclusions Boolean signifying whether excluded points are
@@ -204,6 +208,7 @@ autospc <- function(data,
                     no_regrets = TRUE,
                     overhanging_reversions = TRUE,
                     ## SPC Parameters
+                    na_ends_run = TRUE,
                     max_exclusions = 3L,
                     highlight_exclusions = TRUE,
                     mr_screen_max_loops = 1L,